fix(smtppool): decrement connection count on failed creation
When `p.newConn` returns an error, the number of `createdConns` should
be decremented as well. Otherwise, it could lead to reaching a max
connection count, even if no connections are actually created.

This prevents the pool from falsely believing it is full, which previously led to persistent "timed out waiting for free conn in pool" errors after an SMTP server restart